x/lottery/keeper/query_active_lottery.go
package keeper

import (
	"context"

	"cosmos-lottery/x/lottery/types"
	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) ActiveLottery(goCtx context.Context, req *types.QueryGetActiveLotteryRequest) (*types.QueryGetActiveLotteryResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(goCtx)

	val, found := k.GetActiveLottery(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetActiveLotteryResponse{ActiveLottery: val}, nil
}
